refactor(dev-core): split workspace.ts into 3 modules + concrete resolveCurrentProject
Closes the two deferred items from PR #100 round-3 review:

A1 (warning): split cli/lib/workspace.ts (193 lines, 3 orthogonal concerns)
into:
  - workspace-store.ts   — types + getWorkspacePath/read/write
  - cwd-resolver.ts      — parseGitRemoteUrl/resolveRepoFromCwd/
                            resolveCurrentProject/detectLocalPath + constants
  - github-discovery.ts  — discoverProject (GraphQL)

workspace.ts becomes a barrel re-export so existing importers
(commands/issues.ts, commands/workspace.ts, __tests__/*) work unchanged.

A4 (suggestion): replace resolveCurrentProject's `<P extends ...>`
generic with concrete `WorkspaceProject` parameter. The single non-test
caller (commands/issues.ts) duplicated WorkspaceProject as
IssuesCommandProject with subtly different optional fields and an
`id?` fallback that was dead code (discoverProject always writes
`projectId`, never `id`). Deleted the duplicate, dropped the fallback,
use WorkspaceProject directly.

Tests: 48 pass / 5 fail — same baseline as before (the 5 failures are
pre-existing in issues.test.ts, unrelated to this change).
